refactor(tui): merge Dispatcher into App
Collapses the App / Dispatcher two-type split into a single `App`
struct. Dispatcher was extracted in #212 Phase 1 as "the GoF Receiver
for UserCommand", but every feature added since landed on the state
side, turning Dispatcher into a god-struct and reducing App to a thin
router. Similar TUI projects (Helix, gitui, Yazi, Slumber, …) all
use a single state hub with the loop as a method on it; ttymap's
split was an outlier.

Merging restores that conventional shape. No functional changes —
the previous refactors' invariants are preserved:

- Single bus.publish call site (App::publish_pending drains
  pending_events).
- Four entry points for state mutation: dispatch, accept_frame,
  handle_input, forward_external_event.
- Inline event push beside the state mutation that produced it.
- Lua-agnostic EventBus held on App; LuaHandle holds it only to
  fan tick out.

Doc updates: CLAUDE.md src tree + App module description; lib.rs
module doc; docs/lua-architecture.md drain_ops paragraph + handle.rs
description; lua/mod.rs LuaSubsystem.bus field comment.

Net −86 lines.
